rafactor: move ProfilePage component
