+++ src/components/ProfilePage/ProfilePage.js
@@ -1,6 +1,6 @@
-import Header from '../../common/Header/Header';
-import Navigation from '../../common/Navigation/Navigation';
-import Profile from '../../Profile/Profile';
+import Header from '../common/Header/Header';
+import Navigation from '../common/Navigation/Navigation';
+import Profile from '../Profile/Profile';
 
 function ProfilePage() {
   return (
